add svn:ignore list
git-svn-id: http://svn.cern.ch/guest/lhcb/Rec/trunk@119045 4525493e-7705-40b1-a816-d608a930855b

Former-commit-id: cc557b3aad23411c069206c3d230d8be56dec98e

+++ Rec/Rich/RichRecQC/src/RichRecoQC.cpp
@@ -375,10 +375,10 @@ StatusCode RecoQC::execute()
 
     } // photon loop
 
-    richHisto1D(HID("totalPhotonsPerSeg",rad))->fill(segment->richRecPhotons().size());
+    richHisto1D(HID("totalPhotonsPerSeg",rad))->fill((double)segment->richRecPhotons().size());
     if ( isolated )
     {
-      richHisto1D(HID("totalPhotonsPerSegIso",rad))->fill(segment->richRecPhotons().size());
+      richHisto1D(HID("totalPhotonsPerSegIso",rad))->fill((double)segment->richRecPhotons().size());
     }
 
     // number of true photons
